pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "refute"
version = "0.1.0"
dependencies = [
    "datasets>=3.0.0",
    "langchain_core>=0.3.0",
    "langchain_google_genai>=2.0.0",
    "langchain_openai>=0.3.0",
    "psutil>=6.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0"
]

[tool.setuptools]
packages = ["refute"]